refactor: types 폴더 구조 변경
User, Folder, Link 타입을 api 폴더 내에 배치
