+++ utils/api/types.ts
@@ -1,0 +1,24 @@
+export interface Folder {
+  id: number;
+  created_at: string;
+  name: string;
+  user_id: number;
+}
+
+export interface Link {
+  id: number;
+  created_at: string;
+  updated_at: string | null;
+  url: string;
+  title: string;
+  description: string;
+  image_source: string | null;
+}
+
+export interface User {
+  id: number;
+  created_at: string;
+  name: string;
+  image_source: string;
+  email: string;
+}
